third_party/WebKit/LayoutTests/editing/deleting/delete_br_after_block.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>foo</div>',
            '|<br>',
            'bar',
        '</div>',
    ].join(''),
    'delete',
    [
        '<div contenteditable>',
            '<div>foo|</div>',
            'bar',
        '</div>',
    ].join('')),
    '1 Delete onthe BR after the block');

test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>f^oo</div>',
            '|<br>',
            'bar',
        '</div>',
    ].join(''),
    'delete',
    [
        '<div contenteditable>',
            '<div>f|</div>',
            'bar',
        '</div>',
    ].join('')),
    '2 Delete the whole selection and not just the trailing BR');
</script>
